Context information in the right side (prism) for alternative rule
When the module has a context information, the XOR usually used in alternative patterns is now replaced by CTX. The context is no longer a guard condition but a probability (1 or 0) in which the next state will be running or skipped. CTX represents the context informations of the task. It's an int and should be valued to 0 or 1.